// src/app/sysrev/models/BonSortie.ts
export class BonSortie{
    codeOffice: string = "";
    office: string = "";
    numeroBonSortie: string = "";
    refLiquidation: string = "";
    numeroEnregistrement: string = "";
    numeroQuittance: string = "";
    credit: string = "";
    codeDeclarant: string = "";
    codeEntreprise: string = "";
    dateSortie: string = "";
}